fix: widen RangeCheck fields to i128 so range checks compile on wasm32
The emitted i64-window width guard for control/range-constrained signed
`int` fields (two-major-type signed path) formats `i64::MIN`/`i64::MAX`
as decimal literals into `DeserializeFailure::RangeCheck`'s fields. Those
fields were `isize`, which is 32-bit on `wasm32-unknown-unknown`, so the
literals overflowed and any such generated crate failed to compile for
the real wasm target (host builds are 64-bit and never saw it).

Widen `found`/`min`/`max` to `i128` in the static runtime and at every
emission site (`range_check_err`, the `width_reject` closure, and
`generate_wrapper_struct`). The Display impl formats with `{}` and needs
no change; the bare-literal constructions in `non_empty*.rs` infer.

Pin the class on wasm32: `extern_wrapper_index_defers_to_dep` is the only
gate that builds a generated crate for wasm32-unknown-unknown. Its fixture
gains `local_thing.c: (int .ne 1)`, a constrained signed int whose guard
emits the i64 window — RED (isize literal overflow) before this change,
GREEN after. Also mirror the widened fields into the shared
`tests/extern-dep-crate` runtime the fixture imports.

+++ static/error.rs
@@ -43,9 +43,9 @@ pub enum DeserializeFailure {
     NoVariantMatched,
     NoVariantMatchedWithCauses(Vec<DeserializeError>),
     RangeCheck{
-        found: isize,
-        min: Option<isize>,
-        max: Option<isize>,
+        found: i128,
+        min: Option<i128>,
+        max: Option<i128>,
     },
     /// A float value fell outside its CDDL window (`float64 .le 10.5`, `0.5..10.5`). Separate from
     /// `RangeCheck` because float bounds are f64-typed and carry per-side exclusivity (dense float
